# EventManager/app/docs/swagger.yaml
basePath: /api/event-manager
definitions:
  hateoas.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
      title:
        type: string
    type: object
...
  httpdto.HttpCreateEventPacket:
    properties:
      allocated_seats:
        type: integer
      description:
        type: string
      id_owner:
        type: integer
      location:
        type: string
      name:
        type: string
    required:
    - id_owner
    - name
    type: object
...
  httpdto.HttpResponseEventPacket:
    properties:
      event_packet:
        $ref: '#/definitions/httpdto.httpResponseEventPacket'
    type: object
...
  httpdto.HttpUpdateEventPacket:
    properties:
      allocated_seats:
        type: integer
      description:
        type: string
      id_owner:
        type: integer
      location:
        type: string
      name:
        type: string
    type: object
...
  httpdto.httpResponseEventPacket:
    properties:
      _links:
        additionalProperties:
          $ref: '#/definitions/hateoas.Link'
        type: object
      allocated_seats:
        type: integer
      description:
        type: string
      id:
        type: integer
      id_owner:
        type: integer
      location:
        type: string
      name:
        type: string
    type: object
host: localhost:12345
info:
  contact: {}
  description: Event management service for the POS system. Handles events, event
    packets, tickets, and inclusions.
  title: EventManager Service API
  version: "1.0"
paths:
...
  /event-packets:
    post:
      consumes:
      - application/json
      description: Create a new event packet (ticket package/bundle)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event packet details
        in: body
        name: packet
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpCreateEventPacket'
      produces:
      - application/json
      responses:
        "201":
          description: Event packet created successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacket'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Event packet already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new event packet
      tags:
      - event-packets
  /event-packets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event packet by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Packet ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event packet deleted successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacket'
        "400":
          description: Invalid event packet ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an event packet
      tags:
      - event-packets
    get:
      consumes:
      - application/json
      description: Retrieve a specific event packet by its unique identifier
      parameters:
      - description: Event Packet ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token (optional)
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event packet details
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacket'
        "400":
          description: Invalid event packet ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get event packet by ID
      tags:
      - event-packets
    patch:
      consumes:
      - application/json
      description: Partially update event packet details (PATCH - only provided fields
        are updated)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Packet ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: packet
        required: true
        schema:
          $ref: '#/definitions/httpdto.HttpUpdateEventPacket'
      produces:
      - application/json
      responses:
        "200":
          description: Event packet updated successfully
          schema:
            $ref: '#/definitions/httpdto.HttpResponseEventPacket'
        "400":
          description: Invalid request body or event packet ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Event packet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing event packet
      tags:
      - event-packets
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"